+++ src/Info.php
@@ -10,17 +10,6 @@ use Innmind\Immutable\Maybe;
  */
 final class Info
 {
-    /** @var Maybe<Signal\Code> */
-    private Maybe $code;
-    /** @var Maybe<Signal\ErrorNumber> */
-    private Maybe $errorNumber;
-    /** @var Maybe<Signal\SendingProcessId> */
-    private Maybe $sendingProcessId;
-    /** @var Maybe<Signal\SendingProcessUserId> */
-    private Maybe $sendingProcessUserId;
-    /** @var Maybe<Signal\Status> */
-    private Maybe $status;
-
     /**
      * @param Maybe<Signal\Code> $code
      * @param Maybe<Signal\ErrorNumber> $errorNumber
@@ -28,18 +17,38 @@ final class Info
      * @param Maybe<Signal\SendingProcessUserId> $sendingProcessUserId
      * @param Maybe<Signal\Status> $status
      */
-    public function __construct(
+    private function __construct(
+        private Maybe $code,
+        private Maybe $errorNumber,
+        private Maybe $sendingProcessId,
+        private Maybe $sendingProcessUserId,
+        private Maybe $status,
+    ) {
+    }
+
+    /**
+     * @internal
+     *
+     * @param Maybe<Signal\Code> $code
+     * @param Maybe<Signal\ErrorNumber> $errorNumber
+     * @param Maybe<Signal\SendingProcessId> $sendingProcessId
+     * @param Maybe<Signal\SendingProcessUserId> $sendingProcessUserId
+     * @param Maybe<Signal\Status> $status
+     */
+    public static function of(
         Maybe $code,
         Maybe $errorNumber,
         Maybe $sendingProcessId,
         Maybe $sendingProcessUserId,
         Maybe $status,
-    ) {
-        $this->code = $code;
-        $this->errorNumber = $errorNumber;
-        $this->sendingProcessId = $sendingProcessId;
-        $this->sendingProcessUserId = $sendingProcessUserId;
-        $this->status = $status;
+    ): self {
+        return new self(
+            $code,
+            $errorNumber,
+            $sendingProcessId,
+            $sendingProcessUserId,
+            $status,
+        );
     }
 
     /**